Refactored oscillators into better inheritance structure, and fixed various violations of PIMPL pattern

+++ Synthesizer/Envelopes/Base Implementations/ADSREnvelope.h
@@ -27,7 +27,7 @@ public:
 
     void on() override;
     void off() override;
-    [[nodiscard]] float get_multiplier() override;
+    [[nodiscard]] double get_multiplier() override;
     [[nodiscard]] const EnvelopeState& get_state() const override;
 
     void set_attack(float seconds);
